+++ addons/yat/src/attributes/CommandInputAttribute.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Collections.Generic;
 using Godot;
+using YAT.Classes;
 using YAT.Helpers;
 using YAT.Types;
 
@@ -10,7 +11,7 @@ namespace YAT.Attributes;
 public class CommandInputAttribute : Attribute
 {
 	public StringName Name { get; private set; }
-	public LinkedList<CommandInputType> Types { get; private set; } = new();
+	public List<CommandInputType> Types { get; private set; } = new();
 	public string Description { get; private set; }
 
 	public CommandInputAttribute(StringName name, string type, string description = "")
@@ -19,14 +20,14 @@ public class CommandInputAttribute : Attribute
 		Description = description;
 
 		if (string.IsNullOrEmpty(type))
-			GD.PushError($"Invalid command input type '{type}' for command '{name}'.");
+			GD.PushError(Messages.InvalidCommandInputType(type, name));
 
 		foreach (var t in type.Split('|'))
 		{
-			if (Text.TryParseCommandInputType(t, out var commandInputType))
-				Types.AddLast(commandInputType);
+			if (Parser.TryParseCommandInputType(t, out var commandInputType))
+				Types.Add(commandInputType);
 			else
-				GD.PushError($"Invalid command input type '{t}' for command '{name}'.");
+				GD.PushError(Messages.InvalidCommandInputType(t, name));
 		}
 	}
 }